% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdx.R
\name{cdx_get_status}
\alias{cdx_get_status}
\title{Get Status}
\usage{
cdx_get_status(session, dataset_id)
}
\arguments{
\item{session}{CDX session object}

\item{dataset_id}{Dataset ID from cdx_import}
}
\value{
Status
}
\description{
Check the dataset import or submission status on CDX session.
}
