% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdx.R
\name{cdx_import}
\alias{cdx_import}
\title{Import file}
\usage{
cdx_import(session, file_id, config_id, params = NULL)
}
\arguments{
\item{session}{CDX session object}

\item{file_id}{file_id from cdx_upload}

\item{config_id}{WQX template configuration ID. Found in Import Configurations
under Setup CDX website.}

\item{params}{Configuration parameters for import. Example: ("newOrExistingData", "0")
Do we want to list all of the available import params?}
}
\value{
dataset_id
}
\description{
Use the file ID from the upload function to import the file
to the staging area in the CDX session.
}
