% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdx.R
\name{cdx_submit_to_cdx}
\alias{cdx_submit_to_cdx}
\title{Submit to CDX}
\usage{
cdx_submit_to_cdx(session, dataset_id)
}
\arguments{
\item{session}{CDX session object}

\item{dataset_id}{Dataset ID from cdx_import}
}
\description{
Submit the imported dataset to CDX.
}
