% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdx.R
\name{cdx_upload}
\alias{cdx_upload}
\title{Upload file}
\usage{
cdx_upload(session)
}
\arguments{
\item{session}{CDX session object}
}
\value{
file_ID
}
\description{
Using created CDX session object to upload the file.
}
