% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{wqx_header}
\alias{wqx_header}
\title{Prepare Header}
\usage{
wqx_header(method, uri, config, verbose = FALSE)
}
\arguments{
\item{method}{the method used for the call "GET", "POST", ...}

\item{uri}{uri for the API call}

\item{config}{config file where userID and private key are set. See details section
for more information}
}
\value{
string vector of headers
}
\description{
Create the required headers for all API calls
}
