% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring-locations.R
\name{wqx_locations}
\alias{wqx_locations}
\title{WQX Monitoring Locations Endpoint}
\usage{
wqx_locations(org_id, config = NULL, pk = NULL, username = NULL)
}
\arguments{
\item{org_id}{organization ID as designated by the WQX (required)}

\item{config}{path to configuration file (see details for more information)}

\item{location_id}{a monitoring location ID as designation by organization (optional)}

\item{location_name}{a monitoring location name as designated by organization (optional)}

\item{location_type}{a monitoring location type (optional)}
}
\description{
WQX Monitoring Locations Endpoint
}
