% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{wqx_projects}
\alias{wqx_projects}
\title{Upload to Server}
\usage{
wqx_projects(
  org_id = NULL,
  project_id = NULL,
  transaction_id = NULL,
  last_change_min = NULL,
  last_change_max = NULL,
  ...
)
}
\arguments{
\item{file_path}{path to file, filename cannot have spaces}

\item{private_key}{private key associated with application or user, provided by WQX}

\item{user_id}{user id associated with user or application}
}
\description{
upload a file to wqx server. This does not import the file, rather
keeps it in a staging state.
}
