% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{wqx_start_import}
\alias{wqx_start_import}
\title{Start importing into the staging database}
\usage{
wqx_start_import(
  file_id,
  import_config,
  file_type,
  new_or_existing,
  headers,
  config = NULL,
  pk = NULL,
  username = NULL
)
}
\arguments{
\item{file_id}{file_id for uploaded file, this is returned by wqx_upload()}

\item{import_config}{config id for the destination of file}

\item{file_type}{file type of dataset "CSV"}

\item{new_or_existing}{is this data 'new' data, 'existing' data (replaces old data) or 'both'}

\item{headers}{is the first line of the file a header?}

\item{config}{config file}

\item{pk}{private key if not supplied via a config file}

\item{username}{username if not uspplied via a config file}
}
\description{
import an uploaded dataset into a staging phase
}
