% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{wqx_upload}
\alias{wqx_upload}
\title{Upload to Server}
\usage{
wqx_upload(file_path, private_key, user_id)
}
\arguments{
\item{file_path}{path to file, filename cannot have spaces}

\item{private_key}{private key associated with application or user, provided by WQX}

\item{user_id}{user id associated with user or application}
}
\description{
upload a file to wqx server. This does not import the file, rather
keeps it in a staging state.
}
