% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{wqx_upload}
\alias{wqx_upload}
\title{Upload a file to the WQX (staging)}
\usage{
wqx_upload(file, config = NULL, pk = NULL, username = NULL)
}
\arguments{
\item{file}{file to be uploaded}

\item{config}{config file with private key and username}

\item{pk}{private key if not supplied via a config file}

\item{username}{username if not uspplied via a config file}
}
\value{
returns a list with status code and if succesful a file-id or if not
successful an error message.
}
\description{
upload a file to the wqx server, this file is not pushed to staging yet.
}
